+++ kwcoco/coco_dataset.py
@@ -1390,13 +1390,63 @@ class MixinCocoExtras(object):
         Called whenever the coco dataset is modified. It is possible to specify
         which parts were modified so unmodified parts can be reused next time
         the hash is constructed.
+
+        TODO:
+            - [ ] Rename to _notify_modification --- or something like that
         """
+        self._state['was_modified'] = True
         self.hashid = None
         if parts is not None and self.hashid_parts is not None:
             for part in parts:
                 self.hashid_parts.pop(part, None)
         else:
             self.hashid_parts = None
+
+    def _cached_hashid(self):
+        """
+        Under Construction.
+        """
+        cache_miss = True
+        enable_cache = (
+            self._state['was_loaded'] and
+            not self._state['was_modified']
+        )
+        if enable_cache:
+            coco_fpath = ub.Path(self.fpath)
+            enable_cache = coco_fpath.exists()
+
+        if enable_cache:
+            hashid_sidecar_fpath = ub.Path(str(coco_fpath) + '.hashid.cache')
+            # Generate current lookup key
+            fpath_stat = coco_fpath.stat()
+            status_key = {
+                'st_size': fpath_stat.st_size,
+                'st_mtime': fpath_stat.st_mtime
+            }
+            if hashid_sidecar_fpath.exists():
+                cached_data = json.loads(hashid_sidecar_fpath.read_text())
+                if cached_data['status_key'] == status_key:
+                    self.hashid = cached_data['hashid']
+                    self.hashid_parts = cached_data['hashid_parts']
+                    cache_miss = False
+
+        if cache_miss:
+            self._build_hashid()
+            hashid = self.hashid
+            hashid_parts = self.hashid_parts
+            hashid_cache_data = {
+                'hashid': hashid,
+                'hashid_parts': hashid_parts,
+                'status_key': status_key,
+            }
+            hashid_sidecar_fpath.write_text(json.dumps(hashid_cache_data))
+        return self.hashid
+
+    def _dataset_id(self):
+        hashid = self._cached_hashid()
+        coco_fpath = ub.Path(self.fpath)
+        dset_id = '_'.join([coco_fpath.parent.stem, coco_fpath.stem, hashid[0:8]])
+        return dset_id
 
     def _ensure_imgsize(self, workers=0, verbose=1, fail=False):
         """
@@ -1755,8 +1805,7 @@ class MixinCocoExtras(object):
                new_prefix=None,
                absolute=False,
                check=True,
-               safe=True,
-               smart=False):
+               safe=True):
         """
         Modify the prefix of the image/data paths onto a new image/data root.
 
@@ -1786,11 +1835,6 @@ class MixinCocoExtras(object):
 
             safe (bool, default=True):
                 if True, does not overwrite values until all checks pass
-
-            smart (bool, default=False):
-                If True, we can try different reroot strategies and choose the
-                one that works. Note, always be wary when algorithms try to be
-                smart. NOT IMPLEMENTED. DEPRECATE or TODO?
 
         CommandLine:
             xdoctest -m kwcoco.coco_dataset MixinCocoExtras.reroot
@@ -1848,9 +1892,6 @@ class MixinCocoExtras(object):
         if new_img_root is None:
             new_img_root = self.bundle_dpath
 
-        if smart:
-            raise NotImplementedError('we are not smart yet (probably a good thing)')
-
         def _reroot_path(file_name):
             """ Reroot a single file """
 
@@ -1872,6 +1913,8 @@ class MixinCocoExtras(object):
             # respect to this new root (assuming the previous root was also
             # valid)
             if old_prefix is None and new_prefix is None:
+                # This is not a good check, fails if we want to
+                # do a relative reroot outside of the original dataset
                 if new_img_root is not None and cur_gpath.startswith(new_img_root):
                     file_name = relpath(cur_gpath, new_img_root)
 
@@ -4606,7 +4649,7 @@ class CocoDataset(AbstractCocoDataset, MixinCocoAddRemove, MixinCocoStats,
         >>> # Start by looking up an image by it's COCO id.
         >>> image_id = 1
         >>> img = self.index.imgs[image_id]
-        >>> print(ub.repr2(img, nl=1))
+        >>> print(ub.repr2(img, nl=1, sort=1))
         {
             'file_name': 'astro.png',
             'id': 1,
@@ -4628,7 +4671,7 @@ class CocoDataset(AbstractCocoDataset, MixinCocoAddRemove, MixinCocoStats,
         >>> # Use annotation category id to look up that information
         >>> category_id = ann['category_id']
         >>> cat = self.index.cats[category_id]
-        >>> print('cat = {}'.format(ub.repr2(cat, nl=1)))
+        >>> print('cat = {}'.format(ub.repr2(cat, nl=1, sort=1)))
         cat = {
             'id': 1,
             'name': 'astronaut',
@@ -4637,7 +4680,7 @@ class CocoDataset(AbstractCocoDataset, MixinCocoAddRemove, MixinCocoStats,
         >>> #
         >>> # Now play with some helper functions, like extended statistics
         >>> extended_stats = self.extended_stats()
-        >>> print('extended_stats = {}'.format(ub.repr2(extended_stats, nl=1, precision=2)))
+        >>> print('extended_stats = {}'.format(ub.repr2(extended_stats, nl=1, precision=2, sort=1)))
         extended_stats = {
             'annots_per_img': {'mean': 3.67, 'std': 3.86, 'min': 0.00, 'max': 9.00, 'nMin': 1, 'nMax': 1, 'shape': (3,)},
             'imgs_per_cat': {'mean': 0.88, 'std': 0.60, 'min': 0.00, 'max': 2.00, 'nMin': 2, 'nMax': 1, 'shape': (8,)},
@@ -4691,6 +4734,14 @@ class CocoDataset(AbstractCocoDataset, MixinCocoAddRemove, MixinCocoStats,
                 deprecated alias for bundle_dpath
         """
         self._fpath = None
+
+        # Info about what was the origin of this object and if anything
+        # happened to it over its lifetime.
+        self._state = {
+            'was_loaded': False,
+            'was_saved': False,
+            'was_modified': 0,
+        }
 
         if img_root is not None:
             bundle_dpath = img_root
@@ -4781,9 +4832,11 @@ class CocoDataset(AbstractCocoDataset, MixinCocoAddRemove, MixinCocoStats,
                     '''
                     Specified fpath={} does not exist. If you are trying
                     to create a new dataset fist create a CocoDataset without
-                    any arguments, and then set the fpath attribute
+                    any arguments, and then set the fpath attribute.
+                    We may loosen this requirement in the future.
                     ''').format(fpath))
 
+            self._state['was_loaded'] = True
             with open(fpath, 'r') as file:
                 data = json.load(file)
 
@@ -5119,6 +5172,8 @@ class CocoDataset(AbstractCocoDataset, MixinCocoAddRemove, MixinCocoStats,
                     print('Failed to dump ex = {!r}'.format(ex))
                     self._check_json_serializable()
                     raise
+                else:
+                    self._state['was_saved'] = True
 
     def _check_json_serializable(self, verbose=1):
         """