+++ CHANGELOG.md
@@ -5,7 +5,15 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.2.20 - Unreleased
+## Version 0.2.21 - Unreleased
+
+
+### Changed
+* Minor improvements to confusion measures.
+* AP / ROC curves now report real/support using concise SI for large numbers
+
+
+## Version 0.2.20 - Released 2022-01-18
 
 ### Changed